+++ package.json
@@ -9,7 +9,8 @@
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.4"
+    "react-scripts": "1.1.4",
+    "yup": "^0.25.1"
   },
   "scripts": {
     "start": "react-scripts start",
