+++ src/components/Account.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import AuthUserContext from './AuthUserContext'
-import { PasswordForgetForm } from './PasswordForget'
-import PasswordChangeForm from './PasswordChange'
+import PasswordForgetForm from './PasswordForgetForm'
+import PasswordChangeForm from './PasswordChangeForm'
 import withAuthorization from './withAuthorization'
 
 const AccountPage = () => (
