+++ src/components/NavigationAuth.js
@@ -1,0 +1,27 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+import SignOutButton from './SignOut'
+
+import * as routes from '../constants/routes'
+
+const NavigationAuth = () => (
+  <ul>
+    <li>
+      <Link to={routes.SIGN_IN}>Sign In</Link>
+    </li>
+    <li>
+      <Link to={routes.LANDING}>Landing</Link>
+    </li>
+    <li>
+      <Link to={routes.HOME}>Home</Link>
+    </li>
+    <li>
+      <Link to={routes.ACCOUNT}>Account</Link>
+    </li>
+    <li>
+      <SignOutButton />
+    </li>
+  </ul>
+)
+
+export default NavigationAuth
