+++ src/components/PasswordForget.js
@@ -1,7 +1,5 @@
-import React, { Component } from 'react'
-import { Link } from 'react-router-dom'
-
-import { auth } from '../firebase'
+import React from 'react'
+import PasswordForgetForm from './PasswordForgetForm'
 
 const PasswordForgetPage = () => (
   <div>
@@ -10,65 +8,4 @@ const PasswordForgetPage = () => (
   </div>
 )
 
-const byPropKey = (propertyName, value) => () => ({
-  [propertyName]: value,
-})
-
-const INITIAL_STATE = {
-  email: '',
-  error: null,
-}
-
-class PasswordForgetForm extends Component {
-  constructor(props) {
-    super(props)
-    this.state = { ...INITIAL_STATE }
-  }
-
-  onSubmit(event) {
-    const { email } = this.state
-
-    auth
-      .doPasswordReset(email)
-      .then(() => {
-        this.setState(() => ({ ...INITIAL_STATE }))
-      })
-      .catch(error => {
-        this.setState(byPropKey('error', error))
-      })
-
-    event.preventDefault()
-  }
-
-  render() {
-    const { email, error } = this.state
-
-    const isInvalid = email === ''
-
-    return (
-      <form onSubmit={this.onSubmit.bind(this)}>
-        <input
-          onChange={event => this.setState(byPropKey('email', event.target.value))}
-          placeholder="Email Address"
-          type="text"
-          value={this.state.email}
-        />
-        <button disabled={isInvalid} type="submit">
-          Reset My Password
-        </button>
-
-        {error && <p>{error.message}</p>}
-      </form>
-    )
-  }
-}
-
-const PasswordForgetLink = () => (
-  <p>
-    <Link to="/pw-forget">Forgot Password?</Link>
-  </p>
-)
-
 export default PasswordForgetPage
-
-export { PasswordForgetForm, PasswordForgetLink }
