+++ src/components/PasswordForgetLink.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+const PasswordForgetLink = () => (
+  <p>
+    <Link to="/pw-forget">Forgot Password?</Link>
+  </p>
+)
+
+export default PasswordForgetLink
