+++ src/components/withAuthentication.js
@@ -5,13 +5,7 @@ import AuthUserContext from './AuthUserContext'
 
 const withAuthentication = Component =>
   class WithAuthentication extends React.Component {
-    constructor(props) {
-      super(props)
-
-      this.state = {
-        authUser: null,
-      }
-    }
+    state = { authUser: null }
 
     componentDidMount() {
       auth.onAuthStateChanged(authUser => {
